requirements.txt
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
discord-webhook==1.3.1
filetype==1.2.0
idna==3.10
imageio==2.36.0
inflate64==1.0.0
lazy_loader==0.4
lxml==5.3.0
multivolumefile==0.2.3
networkx==3.4.1
numpy==2.1.2
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
psutil==6.0.0
py7zr==0.22.0
pybcj==1.0.2
pycryptodomex==3.21.0
pyppmd==1.1.0
PyWavelets==1.7.0
pyzstd==0.16.2
rarfile==4.2
regex==2024.9.11
requests==2.32.3
scandir==1.10.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
soupsieve==2.6
texttable==1.7.0
tifffile==2024.9.20
titlecase==2.4.1
Unidecode==1.3.8
urllib3==2.2.3
watchdog==5.0.3
